requirement.txt
huaweicloudsdkecs
huaweicloudsdkims
huaweicloudsdksmn
huaweicloudsdksms
requests
python-dotenv
openpyxl
pandas